Extracted APIs to separate files
