+++ src/@types.d.ts
@@ -1,10 +1,8 @@
-import type {
-  InfiniteScrollerInitResponse, InfiniteScrollerRequest,
-  InfiniteScrollerResponse
-} from './renderer/src/components/InfiniteScroller';
 import type { SongViewProps } from './renderer/src/components/song/SongView';
-import type { SearchQuery, SearchQuerySuccess } from './main/lib/search-parser/@search-types';
-import { ConfigError, ConfigItem, ConfigSuccess } from './main/lib/template-parser/parser/TemplateParser';
+import type { SearchQuerySuccess } from './main/lib/search-parser/@search-types';
+import { ConfigItem } from './main/lib/template-parser/parser/TemplateParser';
+import { RequestAPI } from './RequestAPI';
+import { ListenAPI } from './ListenAPI';
 
 
 
@@ -175,42 +173,6 @@ export type OsuSearchAbleProperties = "bpm" | "artist" | "creator" | "length" | 
 
 
 
-export type RequestAPI = {
-  "resource::get": (id: ResourceID, table: ResourceTables) => Optional<Song | AudioSource | ImageSource>,
-  "resource::getPath": (id: any) => Result<string, string>,
-
-  "queue::exists": () => boolean,
-  "queue::current": () => Optional<Song>,
-  "queue::next": () => void,
-  "queue::previous": () => void,
-  "queue::play": (song: ResourceID) => void,
-  "queue::playNext": (song: ResourceID) => void,
-  "queue::place": (what: ResourceID, after: ResourceID | undefined) => void,
-  "queue::removeSong": (song: ResourceID | undefined) => void,
-  "queue::create": (payload: QueueCreatePayload) => void,
-  "queue::shuffle": () => void,
-
-  "dir::select": () => Optional<string>,
-  "dir::submit": (dir: string) => void,
-
-  "error::dismissed": () => void,
-
-  "parse::search": (query: string) => SearchQuery,
-  "parse::template": (template: string) => ConfigSuccess | ConfigError,
-
-  "settings::write": <K extends keyof Settings>(key: K, value: any) => void,
-  "settings::get": <K extends keyof Settings>(key: K) => Optional<any>,
-
-  "query::songsPool::init": (payload: SongsQueryPayload) => InfiniteScrollerInitResponse,
-  "query::songsPool": (request: InfiniteScrollerRequest, payload: SongsQueryPayload) => InfiniteScrollerResponse<Song>,
-  "query::queue::init": () => InfiniteScrollerInitResponse,
-  "query::queue": (request: InfiniteScrollerRequest) => InfiniteScrollerResponse<Song>,
-
-  "save::localVolume": (volume: number, song: ResourceID) => void,
-}
-
-
-
 export type LoadingSceneUpdate = {
   current: number,
   hint?: string,
@@ -218,17 +180,3 @@ export type LoadingSceneUpdate = {
 }
 
 export type Scenes = "" | "loading" | "dir-select" | "main" | "error";
-
-export type ListenAPI = {
-  changeScene: (scene: Scenes) => void,
-
-  "loadingScene::setTitle": (title: string) => void,
-  "loadingScene::update": (update: LoadingSceneUpdate) => void,
-
-  "error::setMessage": (msg: string) => void,
-  "queue::songChanged": (song: Song) => void,
-  "queue::created": () => void,
-  "queue::destroyed": () => void,
-
-  "songView::reset": () => void,
-}
